Solved

+++ My Library/Seive.cpp
@@ -11,21 +11,21 @@
 using namespace std;
 ///seive
 bool vis[limit];
-vector<int>prime;
+vector<ll>prime;
 void seive()
 {
     vis[0]=vis[1]=1;
-    for(int i=4; i<limit; i+=2) vis[i] = 1;
-    for(int i=3; i*i<limit; i+=2)
+    for(ll i=4; i<limit; i+=2) vis[i] = 1;
+    for(ll i=3; i*i<limit; i+=2)
     {
         if(vis[i]) continue;
-        for(int j=i*i; j<limit; j+=2*i)
+        for(ll j=i*i; j<limit; j+=2*i)
         {
             vis[j] = 1;
         }
     }
     prime.pb(2);
-    for(int j=3; j<limit; j+=2)
+    for(ll j=3; j<limit; j+=2)
         if(vis[j]==0)
             prime.pb(j);
 }
@@ -36,7 +36,7 @@ int  main()
 //    Freed
 //    Fout
     seive();
-    for(int i=0; i<prime.size(); i++)
+    for(ll i=0; i<prime.size(); i++)
         cout <<prime[i]<<" \n";
     return 0;
 }
